Revert "util: move weaponType"
This reverts commit c1076a159e681995c0096255867c45be799ab11a.
